refactor: use relative base url
Since the app FE is now served from the same directory as the BE, they
shared the same port number and thus wouldn't required absolute url for
api calling. The pros of this approach is also that when the port number
is updated, it only needs to be updated in one place.